[NG23-42] Add new lesson review liveview
