+++ lib/oli_web/live/delivery/student/lesson_live.ex
@@ -1,7 +1,20 @@
 defmodule OliWeb.Delivery.Student.LessonLive do
   use OliWeb, :live_view
 
-  import Ecto.Query, warn: false
+  import OliWeb.Delivery.Student.Utils,
+    only: [page_header: 1, star_icon: 1, scripts: 1]
+
+  alias Oli.Delivery.Attempts.Core.ResourceAttempt
+  alias Oli.Delivery.Attempts.PageLifecycle
+  alias Oli.Delivery.Attempts.PageLifecycle.FinalizationSummary
+  alias Oli.Delivery.Page.PageContext
+  alias Oli.Delivery.{Sections, Settings}
+  alias OliWeb.Common.FormatDateTime
+  alias OliWeb.Components.Delivery.Layouts
+  alias OliWeb.Components.Modal
+  alias OliWeb.Delivery.Student.Utils
+
+  require Logger
 
   on_mount {OliWeb.LiveSessionPlugs.InitPage, :page_context}
   on_mount {OliWeb.LiveSessionPlugs.InitPage, :previous_next_index}
@@ -77,6 +90,60 @@ defmodule OliWeb.Delivery.Student.LessonLive do
     {:noreply, assign(socket, show_loader?: false)}
   end
 
+  def handle_event(
+        "finalize_attempt",
+        _params,
+        %{
+          assigns: %{
+            section: section,
+            page_context: page_context,
+            datashop_session_id: datashop_session_id
+          }
+        } = socket
+      ) do
+    revision_slug = page_context.page.slug
+    attempt_guid = hd(page_context.resource_attempts).attempt_guid
+
+    case PageLifecycle.finalize(section.slug, attempt_guid, datashop_session_id) do
+      {:ok,
+       %FinalizationSummary{
+         graded: true,
+         resource_access: %Oli.Delivery.Attempts.Core.ResourceAccess{id: id},
+         effective_settings: effective_settings
+       }} ->
+        # graded resource finalization success
+        section = Sections.get_section_by(slug: section.slug)
+
+        if section.grade_passback_enabled,
+          do: PageLifecycle.GradeUpdateWorker.create(section.id, id, :inline)
+
+        redirect_to =
+          case effective_settings.review_submission do
+            :allow ->
+              ~p"/sections/#{section.slug}/lesson/#{revision_slug}/attempt/#{attempt_guid}/review"
+
+            _ ->
+              ~p"/sections/#{section.slug}/lesson/#{revision_slug}"
+          end
+
+        {:noreply, redirect(socket, to: redirect_to)}
+
+      {:ok, %FinalizationSummary{graded: false}} ->
+        {:noreply, redirect(socket, to: ~p"/sections/#{section.slug}/lesson/#{revision_slug}")}
+
+      {:error, {reason}}
+      when reason in [:already_submitted, :active_attempt_present, :no_more_attempts] ->
+        {:noreply, put_flash(socket, :error, "Unable to finalize page")}
+
+      e ->
+        error_msg = Kernel.inspect(e)
+        Logger.error("Page finalization error encountered: #{error_msg}")
+        Oli.Utils.Appsignal.capture_error(error_msg)
+
+        {:noreply, put_flash(socket, :error, "Unable to finalize page")}
+    end
+  end
+
   def render(%{view: :practice_page} = assigns) do
     # For practice page the activity scripts and activity_bridge script are needed as soon as the page loads.
     ~H"""
@@ -87,7 +154,7 @@ defmodule OliWeb.Delivery.Student.LessonLive do
             page_context={@page_context}
             ctx={@ctx}
             index={@current_page["index"]}
-            container_label={get_container_label(@current_page["id"], @section)}
+            container_label={Utils.get_container_label(@current_page["id"], @section)}
           />
           <div id="eventIntercept" class="content" phx-update="ignore" role="page content">
             <%= raw(@html) %>
@@ -107,12 +174,7 @@ defmodule OliWeb.Delivery.Student.LessonLive do
           begin_attempt?: false
         } = assigns
       ) do
-    # For graded page with attempt in progrress the activity scripts and activity_bridge script are needed as soon as the page loads.
-
-    # TODO after submitting answers, user should be redirected to the new NG23 review page.
-    # (now it is being redirected to old version of review page)
-    # We should replace the onClick with its equivalent function in Elixir: Oli.Delivery.Attempts.PageLifecycle.finalize
-    # and then redirect to the new review page.
+    # For graded page with attempt in progress the activity scripts and activity_bridge script are needed as soon as the page loads.
     ~H"""
     <div class="flex pb-20 flex-col items-center gap-15 flex-1">
       <div class="flex flex-col items-center w-full">
@@ -122,14 +184,14 @@ defmodule OliWeb.Delivery.Student.LessonLive do
             page_context={@page_context}
             ctx={@ctx}
             index={@current_page["index"]}
-            container_label={get_container_label(@current_page["id"], @section)}
+            container_label={Utils.get_container_label(@current_page["id"], @section)}
           />
           <div id="eventIntercept" class="content" phx-update="ignore" role="page content">
             <%= raw(@html) %>
             <div class="flex w-full justify-center">
               <button
                 id="submit_answers"
-                onClick={"window.OLI.finalize('#{@section.slug}', '#{@page_context.page.slug}', '#{hd(@page_context.resource_attempts).attempt_guid || nil}', #{@page_context.page.graded}, 'submit_answers')"}
+                phx-click="finalize_attempt"
                 class="cursor-pointer px-5 py-2.5 hover:bg-opacity-40 bg-blue-600 rounded-[3px] shadow justify-center items-center gap-2.5 inline-flex text-white text-sm font-normal font-['Open Sans'] leading-tight"
               >
                 Submit Answers
@@ -147,12 +209,6 @@ defmodule OliWeb.Delivery.Student.LessonLive do
   def render(%{view: :graded_page, begin_attempt?: true} = assigns) do
     # For graded page with no started attempts, the js scripts are needed after the user clicks "Begin",
     # so we load them with the hook "load_survey_scripts" in the click handle_event functions.
-
-    # TODO after submitting answers, user should be redirected to the new NG23 review page.
-    # (now it is being redirected to old version of review page)
-    # We should replace the onClick with its equivalent function in Elixir: Oli.Delivery.Attempts.PageLifecycle.finalize
-    # and then redirect to the new review page.
-
     ~H"""
     <div class="flex pb-20 flex-col items-center gap-15 flex-1">
       <div class="flex flex-col items-center w-full">
@@ -162,7 +218,7 @@ defmodule OliWeb.Delivery.Student.LessonLive do
             page_context={@page_context}
             ctx={@ctx}
             index={@current_page["index"]}
-            container_label={get_container_label(@current_page["id"], @section)}
+            container_label={Utils.get_container_label(@current_page["id"], @section)}
           />
           <div id="page_content" phx-hook="LoadSurveyScripts">
             <div
@@ -188,7 +244,7 @@ defmodule OliWeb.Delivery.Student.LessonLive do
               <div class="flex w-full justify-center">
                 <button
                   id="submit_answers"
-                  onClick={"window.OLI.finalize('#{@section.slug}', '#{@page_context.page.slug}', '#{hd(@page_context.resource_attempts).attempt_guid || nil}', #{@page_context.page.graded}, 'submit_answers')"}
+                  phx-click="finalize_attempt"
                   class="cursor-pointer px-5 py-2.5 hover:bg-opacity-40 bg-blue-600 rounded-[3px] shadow justify-center items-center gap-2.5 inline-flex text-white text-sm font-normal font-['Open Sans'] leading-tight"
                 >
                   Submit Answers
@@ -227,7 +283,7 @@ defmodule OliWeb.Delivery.Student.LessonLive do
           page_context={@page_context}
           ctx={@ctx}
           index={@current_page["index"]}
-          container_label={get_container_label(@current_page["id"], @section)}
+          container_label={Utils.get_container_label(@current_page["id"], @section)}
         />
         <div class="self-stretch h-[0px] opacity-80 dark:opacity-20 bg-white border border-gray-200">
         </div>
@@ -236,32 +292,10 @@ defmodule OliWeb.Delivery.Student.LessonLive do
           attempt_message={@attempt_message}
           ctx={@ctx}
           allow_attempt?={@allow_attempt?}
+          section_slug={@section.slug}
         />
       </div>
     </div>
-    """
-  end
-
-  # As we implement more scenarios we can add more clauses to this function depending on the :view key.
-  def render(assigns) do
-    ~H"""
-    <div>hi, this is still in dev</div>
-    """
-  end
-
-  attr :scripts, :list
-  attr :user_token, :string
-
-  def scripts(assigns) do
-    ~H"""
-    <script>
-      window.userToken = "<%= @user_token %>";
-    </script>
-    <script>
-      OLI.initActivityBridge('eventIntercept');
-    </script>
-    <script :for={script <- @scripts} type="text/javascript" src={"/js/#{script}"}>
-    </script>
     """
   end
 
@@ -269,6 +303,7 @@ defmodule OliWeb.Delivery.Student.LessonLive do
   attr :page_context, Oli.Delivery.Page.PageContext
   attr :ctx, OliWeb.Common.SessionContext
   attr :allow_attempt?, :boolean
+  attr :section_slug, :string
 
   defp attempts_summary(assigns) do
     ~H"""
@@ -297,6 +332,8 @@ defmodule OliWeb.Delivery.Student.LessonLive do
               |> Enum.with_index(1)
           }
           index={index}
+          section_slug={@section_slug}
+          page_revision_slug={@page_context.page.slug}
           attempt={attempt}
           ctx={@ctx}
           allow_review_submission?={@page_context.effective_settings.review_submission == :allow}
@@ -327,6 +364,8 @@ defmodule OliWeb.Delivery.Student.LessonLive do
   attr :attempt, ResourceAttempt
   attr :ctx, OliWeb.Common.SessionContext
   attr :allow_review_submission?, :boolean
+  attr :section_slug, :string
+  attr :page_revision_slug, :string
 
   defp attempt_summary(assigns) do
     # TODO: add link to review page
@@ -377,9 +416,13 @@ defmodule OliWeb.Delivery.Student.LessonLive do
           :if={@allow_review_submission?}
           class="w-[124px] py-1 justify-end items-center gap-2.5 inline-flex"
         >
-          <div class="cursor-pointer hover:opacity-40 text-blue-500 text-xs font-semibold font-['Open Sans'] uppercase tracking-wide">
-            Review
-          </div>
+          <.link href={
+            ~p"/sections/#{@section_slug}/lesson/#{@page_revision_slug}/attempt/#{@attempt.attempt_guid}/review"
+          }>
+            <div class="cursor-pointer hover:opacity-40 text-blue-500 text-xs font-semibold font-['Open Sans'] uppercase tracking-wide">
+              Review
+            </div>
+          </.link>
         </div>
       </div>
     </div>
@@ -398,182 +441,6 @@ defmodule OliWeb.Delivery.Student.LessonLive do
         You can start or stop at any time, and your progress will be saved. When you submit your answers using the Submit button, it will count as an attempt. So make sure you have answered all the questions before submitting.
       </div>
     </div>
-    """
-  end
-
-  attr :page_context, Oli.Delivery.Page.PageContext
-  attr :ctx, OliWeb.Common.SessionContext
-  attr :index, :string
-  attr :container_label, :string
-  attr :has_assignments?, :boolean
-
-  def page_header(assigns) do
-    ~H"""
-    <div id="page_header" class="flex-col justify-start items-start gap-9 flex w-full">
-      <div class="flex-col justify-start items-start gap-3 flex w-full">
-        <div class="self-stretch flex-col justify-start items-start flex">
-          <div class="self-stretch justify-between items-center inline-flex">
-            <div class="grow shrink basis-0 self-stretch justify-start items-center gap-3 flex">
-              <div
-                role="container label"
-                class="opacity-50 dark:text-white text-sm font-bold font-['Open Sans'] uppercase tracking-wider"
-              >
-                <%= @container_label %>
-              </div>
-
-              <div
-                :if={@page_context.page.graded}
-                class="w-px self-stretch opacity-40 bg-black dark:bg-white"
-              >
-              </div>
-              <div
-                :if={@page_context.page.graded}
-                class="justify-start items-center gap-1.5 flex"
-                role="graded page marker"
-              >
-                <div class="w-[18px] h-[18px] relative">
-                  <.flag_icon />
-                </div>
-                <div class="opacity-50 dark:text-white text-sm font-bold font-['Open Sans'] uppercase tracking-wider">
-                  Graded Page
-                </div>
-              </div>
-            </div>
-            <div
-              :if={@page_context.page.graded}
-              class="px-2 py-1 bg-black bg-opacity-10 dark:bg-white dark:bg-opacity-10 rounded-xl shadow justify-start items-center gap-1 flex"
-              role="assignment marker"
-            >
-              <div class="dark:text-white text-[10px] font-normal font-['Open Sans']">
-                Assignment requirement
-              </div>
-            </div>
-          </div>
-          <div role="page label" class="self-stretch justify-start items-start gap-2.5 inline-flex">
-            <div
-              role="page numbering index"
-              class="opacity-50 dark:text-white text-[38px] font-bold font-['Open Sans']"
-            >
-              <%= @index %>.
-            </div>
-            <div
-              role="page title"
-              class="grow shrink basis-0 dark:text-white text-[38px] font-bold font-['Open Sans']"
-            >
-              <%= @page_context.page.title %>
-            </div>
-          </div>
-        </div>
-        <div class="justify-start items-center gap-3 inline-flex">
-          <div class="opacity-50 justify-start items-center gap-1.5 flex">
-            <div role="page read time" class="justify-end items-center gap-1 flex">
-              <div class="w-[18px] h-[18px] relative opacity-80">
-                <.time_icon />
-              </div>
-              <div class="justify-end items-end gap-0.5 flex">
-                <div class="text-right dark:text-white text-xs font-bold font-['Open Sans'] uppercase tracking-wide">
-                  <%= @page_context.page.duration_minutes %>
-                </div>
-                <div class="dark:text-white text-[9px] font-bold font-['Open Sans'] uppercase tracking-wide">
-                  min
-                </div>
-              </div>
-            </div>
-          </div>
-          <div role="page schedule" class="justify-start items-start gap-1 flex">
-            <div class="opacity-50 dark:text-white text-xs font-normal font-['Open Sans']">Due:</div>
-            <div class="dark:text-white text-xs font-normal font-['Open Sans']">
-              <%= FormatDateTime.to_formatted_datetime(
-                @page_context.effective_settings.end_date,
-                @ctx,
-                "{WDshort} {Mshort} {D}, {YYYY}"
-              ) %>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div
-        :if={@page_context.objectives not in [nil, []]}
-        class="flex-col justify-start items-start gap-3 flex w-full"
-        role="page objectives"
-      >
-        <div class="self-stretch justify-start items-start gap-6 inline-flex">
-          <div class="opacity-80 dark:text-white text-sm font-bold font-['Open Sans'] uppercase tracking-wider">
-            Learning objectives
-          </div>
-          <div class="hidden text-blue-500 text-sm font-semibold font-['Open Sans']">View More</div>
-        </div>
-        <div
-          :for={{objective, index} <- Enum.with_index(@page_context.objectives, 1)}
-          class="self-stretch flex-col justify-start items-start flex"
-          role={"objective #{index}"}
-        >
-          <div class="self-stretch py-1 justify-start items-start inline-flex">
-            <div class="grow shrink basis-0 h-6 justify-start items-start flex">
-              <div class="w-[30px] opacity-40 dark:text-white text-xs font-bold font-['Open Sans'] leading-normal">
-                L<%= index %>
-              </div>
-              <div class="grow shrink basis-0 opacity-80 dark:text-white text-sm font-normal font-['Open Sans'] leading-normal">
-                <%= objective %>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    """
-  end
-
-  def star_icon(assigns) do
-    ~H"""
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-      role="star icon"
-    >
-      <path
-        d="M3.88301 14.0007L4.96634 9.31732L1.33301 6.16732L6.13301 5.75065L7.99967 1.33398L9.86634 5.75065L14.6663 6.16732L11.033 9.31732L12.1163 14.0007L7.99967 11.5173L3.88301 14.0007Z"
-        fill="#0CAF61"
-      />
-    </svg>
-    """
-  end
-
-  def flag_icon(assigns) do
-    ~H"""
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="18"
-      height="18"
-      viewBox="0 0 18 18"
-      fill="none"
-      role="flag icon"
-    >
-      <path d="M3.75 15.75V3H10.5L10.8 4.5H15V12H9.75L9.45 10.5H5.25V15.75H3.75Z" fill="#F68E2E" />
-    </svg>
-    """
-  end
-
-  def time_icon(assigns) do
-    ~H"""
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="18"
-      height="18"
-      viewBox="0 0 18 18"
-      fill="none"
-      role="time icon"
-    >
-      <g opacity="0.8">
-        <path
-          class="fill-black dark:fill-white"
-          d="M11.475 12.525L12.525 11.475L9.75 8.7V5.25H8.25V9.3L11.475 12.525ZM9 16.5C7.9625 16.5 6.9875 16.3031 6.075 15.9094C5.1625 15.5156 4.36875 14.9813 3.69375 14.3063C3.01875 13.6313 2.48438 12.8375 2.09063 11.925C1.69688 11.0125 1.5 10.0375 1.5 9C1.5 7.9625 1.69688 6.9875 2.09063 6.075C2.48438 5.1625 3.01875 4.36875 3.69375 3.69375C4.36875 3.01875 5.1625 2.48438 6.075 2.09063C6.9875 1.69688 7.9625 1.5 9 1.5C10.0375 1.5 11.0125 1.69688 11.925 2.09063C12.8375 2.48438 13.6313 3.01875 14.3063 3.69375C14.9813 4.36875 15.5156 5.1625 15.9094 6.075C16.3031 6.9875 16.5 7.9625 16.5 9C16.5 10.0375 16.3031 11.0125 15.9094 11.925C15.5156 12.8375 14.9813 13.6313 14.3063 14.3063C13.6313 14.9813 12.8375 15.5156 11.925 15.9094C11.0125 16.3031 10.0375 16.5 9 16.5ZM9 15C10.6625 15 12.0781 14.4156 13.2469 13.2469C14.4156 12.0781 15 10.6625 15 9C15 7.3375 14.4156 5.92188 13.2469 4.75313C12.0781 3.58438 10.6625 3 9 3C7.3375 3 5.92188 3.58438 4.75313 4.75313C3.58438 5.92188 3 7.3375 3 9C3 10.6625 3.58438 12.0781 4.75313 13.2469C5.92188 14.4156 7.3375 15 9 15Z"
-        />
-      </g>
-    </svg>
     """
   end
 
@@ -627,43 +494,28 @@ defmodule OliWeb.Delivery.Student.LessonLive do
       _ ->
         # In the case where a gate exists we want to redirect to this page display, which will
         # then pick up the gate and show that feedback to the user
-        redirect(socket,
-          to: Routes.page_delivery_path(socket, :page, section.slug, revision.slug)
-        )
+        {:noreply,
+         redirect(socket,
+           to: Routes.page_delivery_path(socket, :page, section.slug, revision.slug)
+         )}
     end
   end
 
-  defp get_container_label(page_id, section) do
-    section_id = section.id
+  defp assign_html_and_scripts(socket) do
+    socket
+    |> assign_scripts()
+    |> assign_html()
+  end
 
-    # Query to find the parent section_resource which contains as a child
-    # the section resource whose resource_id matches the given page_id. Just
-    # be more robust against weird hierarchies and maybe orphaned containers,
-    # we look for only containers with a numbering_level >= 0 and limit to 1.
+  defp assign_scripts(socket) do
+    assign(socket,
+      scripts: Utils.get_required_activity_scripts(socket.assigns.page_context)
+    )
+  end
 
-    query =
-      from(s in Oli.Delivery.Sections.SectionResource,
-        join: sr in Oli.Delivery.Sections.SectionResource,
-        on: sr.section_id == s.section_id and sr.resource_id == ^page_id,
-        where:
-          s.section_id == ^section_id and
-            sr.id in s.children and
-            s.numbering_level >= 0,
-        select: s,
-        limit: 1
-      )
-
-    # If we find a container, use it to get the label and numbering. Otherwise,
-    # fall back rendering something
-    container =
-      case Oli.Repo.all(query) do
-        [item] -> item
-        [] -> %{numbering_index: 0, numbering_level: 0}
-      end
-
-    Sections.get_container_label_and_numbering(
-      container,
-      section.customizations
+  defp assign_html(socket) do
+    assign(socket,
+      html: Utils.build_html(socket.assigns, :delivery)
     )
   end
 
@@ -691,96 +543,9 @@ defmodule OliWeb.Delivery.Student.LessonLive do
     end
   end
 
-  defp assign_html_and_scripts(socket) do
-    socket
-    |> assign_scripts()
-    |> assign_html()
-  end
-
-  defp assign_scripts(socket) do
-    assign(socket,
-      scripts: get_required_activity_scripts(socket.assigns.page_context)
-    )
-  end
-
   defp load_scripts_on_client_side(socket) do
     push_event(socket, "load_survey_scripts", %{
       script_sources: Enum.map(socket.assigns.scripts, fn script -> "/js/#{script}" end)
     })
-  end
-
-  defp assign_html(socket) do
-    %{section: section, current_user: current_user, page_context: page_context} = socket.assigns
-
-    render_context = %Context{
-      enrollment:
-        Oli.Delivery.Sections.get_enrollment(
-          section.slug,
-          current_user.id
-        ),
-      user: current_user,
-      section_slug: section.slug,
-      mode: :delivery,
-      activity_map: page_context.activities,
-      resource_summary_fn: &Resources.resource_summary(&1, section.slug, Resolver),
-      alternatives_groups_fn: fn ->
-        Resources.alternatives_groups(section.slug, Resolver)
-      end,
-      alternatives_selector_fn: &Resources.Alternatives.select/2,
-      extrinsic_read_section_fn: &Oli.Delivery.ExtrinsicState.read_section/3,
-      bib_app_params: page_context.bib_revisions,
-      historical_attempts: page_context.historical_attempts,
-      learning_language: Sections.get_section_attributes(section).learning_language,
-      effective_settings: page_context.effective_settings
-      # when migrating from page_delivery_controller this key-values were found
-      # to apparently not be used by the page template:
-      #   project_slug: base_project_slug,
-      #   submitted_surveys: submitted_surveys,
-      #   resource_attempt: hd(context.resource_attempts)
-    }
-
-    attempt_content = get_attempt_content(page_context)
-
-    # Cache the page as text to allow the AI agent LV to access it.
-    cache_page_as_text(render_context, attempt_content, page_context.page.id)
-
-    assign(socket,
-      html: Page.render(render_context, attempt_content, Page.Html)
-    )
-  end
-
-  defp cache_page_as_text(render_context, content, page_id) do
-    Oli.Converstation.PageContentCache.put(
-      page_id,
-      Page.render(render_context, content, Page.Markdown) |> :erlang.iolist_to_binary()
-    )
-  end
-
-  defp get_required_activity_scripts(%{activities: activities} = _page_context)
-       when activities != nil do
-    # this is an optimization to exclude not needed activity scripts (~1.5mb each)
-    Enum.map(activities, fn {_activity_id, activity} ->
-      activity.script
-    end)
-    |> Enum.uniq()
-  end
-
-  defp get_required_activity_scripts(_page_context) do
-    # TODO Optimization: get only activity scripts of activities contained in the page.
-    # Could we infer the contained activities from the page revision content model?
-    all_activities = Activities.list_activity_registrations()
-    Enum.map(all_activities, fn a -> a.delivery_script end)
-  end
-
-  defp get_attempt_content(page_context) do
-    this_attempt = page_context.resource_attempts |> hd
-
-    if Enum.any?(this_attempt.errors, fn e ->
-         e == "Selection failed to fulfill: no values provided for expression"
-       end) and page_context.is_student do
-      %{"model" => []}
-    else
-      this_attempt.content
-    end
   end
 end
