+++ lib/oli_web/router.ex
@@ -962,17 +962,20 @@ defmodule OliWeb.Router do
         live("/practice", Delivery.Student.PracticeLive)
       end
 
-      live_session :delivery_lesson,
-        root_layout: {OliWeb.LayoutView, :delivery},
-        layout: {OliWeb.Layouts, :student_delivery_lesson},
-        on_mount: [
-          OliWeb.LiveSessionPlugs.SetUser,
-          OliWeb.LiveSessionPlugs.SetSection,
-          OliWeb.LiveSessionPlugs.SetBrand,
-          OliWeb.LiveSessionPlugs.SetPreviewMode,
-          OliWeb.LiveSessionPlugs.RequireEnrollment
-        ] do
-        live("/lesson/:revision_slug", Delivery.Student.LessonLive)
+      scope "/lesson/:revision_slug" do
+        live_session :delivery_lesson,
+          root_layout: {OliWeb.LayoutView, :delivery},
+          layout: {OliWeb.Layouts, :student_delivery_lesson},
+          on_mount: [
+            OliWeb.LiveSessionPlugs.SetUser,
+            OliWeb.LiveSessionPlugs.SetSection,
+            OliWeb.LiveSessionPlugs.SetBrand,
+            OliWeb.LiveSessionPlugs.SetPreviewMode,
+            OliWeb.LiveSessionPlugs.RequireEnrollment
+          ] do
+          live("/", Delivery.Student.LessonLive)
+          live("/attempt/:attempt_guid/review", Delivery.Student.ReviewLive)
+        end
       end
     end
